+++ .github/workflows/a11y.yml
@@ -1,0 +1,30 @@
+on:
+    push:
+    pull_request:
+        types: [opened, synchronize, reopened]
+name: a11y tests
+env:
+    REDIS_SUPPRESS: true
+jobs:
+    build:
+        name: Test
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v3
+            - uses: actions/setup-node@v3
+              with:
+                  node-version: '18.x'
+                  registry-url: 'https://registry.npmjs.org'
+                  scope: '@octocat'
+            - name: Run install
+              uses: borales/actions-yarn@v4
+              with:
+                  cmd: install # will run `yarn install` command
+            - name: Run webpack
+              uses: borales/actions-yarn@v4
+              with:
+                  cmd: webpack # will run `yarn webpack` command
+            - name: Run test:a11y
+              uses: borales/actions-yarn@v4
+              with:
+                  cmd: test:a11y  # will run `yarn test:a11y` command
