--- src/test/a11y/case-reference-number-search.test.ts
@@ -1,28 +1,0 @@
-import * as supertest from 'supertest';
-import { app } from '../../main/app';
-import { ensurePageCallWillSucceed, expectNoErrors, Pa11yResult, runPally } from './a11y';
-const agent = supertest.agent(app);
-
-const URL = '/case-reference-number-search';
-
-describe('Accessibility Case Reference Number Search Page Error States', () => {
-    test('should have no accessibility errors for no input data', done => {
-        ensurePageCallWillSucceed(URL)
-            .then(() => runPally(agent.post(URL).send({ 'search-input': '' }).url))
-            .then((result: Pa11yResult) => {
-                expectNoErrors(result.issues);
-                done();
-            })
-            .catch((err: Error) => done(err));
-    });
-
-    test('should have no accessibility errors for invalid input data', done => {
-        ensurePageCallWillSucceed(URL)
-            .then(() => runPally(agent.post(URL).send({ 'search-input': '123' }).url))
-            .then((result: Pa11yResult) => {
-                expectNoErrors(result.issues);
-                done();
-            })
-            .catch((err: Error) => done(err));
-    });
-});
